+++ packages/database/src/write-database.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 const { writeJsonGzip, writeSafe } = require('@home-gallery/common');
 
 const wrapEntries = (entries) => {
@@ -11,13 +13,26 @@ const wrapEntries = (entries) => {
 const writeDatabase = (filename, entries, cb) => {
   const database = wrapEntries(entries);
 
-  writeJsonGzip(filename, database, cb);
+  const tmp = `${filename}.tmp`;
+  writeJsonGzip(tmp, database, err => {
+    if (err) {
+      return cb(err);
+    }
+    fs.rename(tmp, filename, err => cb(err, err ? null : database));
+  });
 }
 
 const writeDatabasePlain = (filename, entries, cb) => {
   const database = wrapEntries(entries);
   const data = JSON.stringify(database);
-  writeSafe(filename, data, cb);
+
+  const tmp = `${filename}.tmp`;
+  writeSafe(tmp, data, err => {
+    if (err) {
+      return cb(err);
+    }
+    fs.rename(tmp, filename, err => cb(err, err ? null : database));
+  });
 }
 
 module.exports = { writeDatabase, writeDatabasePlain } ;